requirements.txt
gspread
oath2client
discord.py